pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
[project]
name = 'GoogleCloudPlatformAPI'
version = 'v2.3.5'
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pandas",
    "google-api-python-client",
    "google-cloud-bigquery",
    "google-cloud-bigquery[pandas]",
    "google-cloud-storage",
    "googleads",
    "db-dtypes",
    "python-dotenv",
]

[tool.pydocstyle]
convention = "numpy"
match = '(?!AdManager).*\.py'

[tool.pyright]
exclude = ["GoogleCloudPlatformAPI/AdManager.py"]
